+++ test/unit/git_class/gitconfig_get_email.py
@@ -42,10 +42,6 @@ class UnitTest(unittest.TestCase):
 
     Description:  Class which is a representation of a unit testing.
 
-    Super-Class:  unittest.TestCase
-
-    Sub-Classes:
-
     Methods:
         setUp -> Unit testing initilization.
         test_get_email -> Test get_email method.