+++ test/unit/git_class/gitmerge_create_gitrepo.py
@@ -9,7 +9,6 @@
         test/unit/git_class/gitmerge_create_gitrepo.py
 
     Arguments:
-        None
 
 """
 
@@ -33,7 +32,6 @@ import git_class
 import lib.gen_libs as gen_libs
 import version
 
-# Version
 __version__ = version.__version__
 
 
@@ -42,10 +40,6 @@ class UnitTest(unittest.TestCase):
     """Class:  UnitTest
 
     Description:  Class which is a representation of a unit testing.
-
-    Super-Class:  unittest.TestCase
-
-    Sub-Classes:  None
 
     Methods:
         setUp -> Unit testing initilization.
@@ -60,7 +54,6 @@ class UnitTest(unittest.TestCase):
         Description:  Initialization for unit testing.
 
         Arguments:
-            None
 
         """
 
@@ -82,7 +75,6 @@ class UnitTest(unittest.TestCase):
         Description:  Test with default values settings.
 
         Arguments:
-            None
 
         """