+++ test/unit/git_class/gitmerge_get_untracked.py
@@ -9,7 +9,6 @@
         test/unit/git_class/gitmerge_get_untracked.py
 
     Arguments:
-        None
 
 """
 
@@ -35,7 +34,6 @@ import git_class
 import lib.gen_libs as gen_libs
 import version
 
-# Version
 __version__ = version.__version__
 
 
@@ -57,7 +55,6 @@ class Index(object):
         Description:  Initialization of class instance.
 
         Arguments:
-            None
 
         """
 
@@ -72,6 +69,8 @@ class Diff(Index):
 
     Methods:
         __init -> Class initilization.
+        add -> Method stub holder for git.gitrepo.index.add().
+        commit -> Method stub holder for git.gitrepo.index.commit().
 
     """
 
@@ -82,7 +81,6 @@ class Diff(Index):
         Description:  Initialization of class instance.
 
         Arguments:
-            None
 
         """
 
@@ -121,10 +119,6 @@ class UnitTest(unittest.TestCase):
 
     Description:  Class which is a representation of a unit testing.
 
-    Super-Class:  unittest.TestCase
-
-    Sub-Classes:  None
-
     Methods:
         setUp -> Unit testing initilization.
         test_process_data_list -> Test with data in list set.
@@ -139,7 +133,6 @@ class UnitTest(unittest.TestCase):
         Description:  Initialization for unit testing.
 
         Arguments:
-            None
 
         """
 
@@ -162,7 +155,6 @@ class UnitTest(unittest.TestCase):
         Description:  Test with data in list set.
 
         Arguments:
-            None
 
         """
 
@@ -181,7 +173,6 @@ class UnitTest(unittest.TestCase):
         Description:  Test with empty list set.
 
         Arguments:
-            None
 
         """