+++ test/unit/git_class/gitmerge_is_remote.py
@@ -9,7 +9,6 @@
         test/unit/git_class/gitmerge_is_remote.py
 
     Arguments:
-        None
 
 """
 
@@ -35,7 +34,6 @@ import git_class
 import lib.gen_libs as gen_libs
 import version
 
-# Version
 __version__ = version.__version__
 
 
@@ -73,10 +71,6 @@ class UnitTest(unittest.TestCase):
 
     Description:  Class which is a representation of a unit testing.
 
-    Super-Class:  unittest.TestCase
-
-    Sub-Classes:  None
-
     Methods:
         setUp -> Unit testing initilization.
         test_is_remote_false -> Test with exception raised from ls_remote call.
@@ -91,7 +85,6 @@ class UnitTest(unittest.TestCase):
         Description:  Initialization for unit testing.
 
         Arguments:
-            None
 
         """
 
@@ -111,7 +104,6 @@ class UnitTest(unittest.TestCase):
         Description:  Test with exception raised from ls_remote call.
 
         Arguments:
-            None
 
         """
 
@@ -127,7 +119,6 @@ class UnitTest(unittest.TestCase):
         Description:  Test with successful ls_remote call.
 
         Arguments:
-            None
 
         """