+++ test/unit/git_class/gitmerge_process_untracked.py
@@ -9,7 +9,6 @@
         test/unit/git_class/gitmerge_process_untracked.py
 
     Arguments:
-        None
 
 """
 
@@ -35,7 +34,6 @@ import git_class
 import lib.gen_libs as gen_libs
 import version
 
-# Version
 __version__ = version.__version__
 
 
@@ -57,7 +55,6 @@ class Index(object):
         Description:  Initialization of class instance.
 
         Arguments:
-            None
 
         """
 
@@ -72,6 +69,8 @@ class Diff(Index):
 
     Methods:
         __init -> Class initilization.
+        add -> Method stub holder for git.gitrepo.index.add().
+        commit -> Method stub holder for git.gitrepo.index.commit().
 
     """
 
@@ -82,7 +81,6 @@ class Diff(Index):
         Description:  Initialization of class instance.
 
         Arguments:
-            None
 
         """
 
@@ -121,12 +119,9 @@ class UnitTest(unittest.TestCase):
 
     Description:  Class which is a representation of a unit testing.
 
-    Super-Class:  unittest.TestCase
-
-    Sub-Classes:  None
-
     Methods:
         setUp -> Unit testing initilization.
+        test_process_remove_dir -> 
         test_process_add_option -> Test with add option passed.
         test_process_remove_option -> Test with remove option passed.
         test_process_newfiles_list -> Test with new_files list set.
@@ -142,7 +137,6 @@ class UnitTest(unittest.TestCase):
         Description:  Initialization for unit testing.
 
         Arguments:
-            None
 
         """
 
@@ -158,42 +152,20 @@ class UnitTest(unittest.TestCase):
         self.new_list1 = []
         self.new_list2 = ["file1"]
 
-    @mock.patch("git_class.gen_libs")
-    def test_process_add_option(self, mock_lib):
+    @mock.patch("git_class.os")
+    @mock.patch("git_class.shutil")
+    def test_process_remove_dir(self, mock_shutil, mock_os):
 
-        """Function:  test_process_add_option
+        """Function:  test_process_remove_dir
 
-        Description:  Test with add option passed.
+        Description:  Test with removal of directory.
 
         Arguments:
-            None
 
         """
 
-        mock_lib.rm_file.return_value = True
-
-        GIT = collections.namedtuple('GIT', 'index untracked_files')
-        DIFF = Diff()
-        self.gitr.gitrepo = GIT(DIFF, self.new_list1)
-        self.gitr.new_files = self.new_list2
-
-        self.gitr.process_untracked("add")
-
-        self.assertEqual(self.gitr.new_files, self.new_list2)
-
-    @mock.patch("git_class.gen_libs")
-    def test_process_remove_option(self, mock_lib):
-
-        """Function:  test_process_remove_option
-
-        Description:  Test with remove option passed.
-
-        Arguments:
-            None
-
-        """
-
-        mock_lib.rm_file.return_value = True
+        mock_shutil.rmtree.return_value = True
+        mock_os.path.isdir.return_value = True
 
         GIT = collections.namedtuple('GIT', 'index untracked_files')
         DIFF = Diff()
@@ -204,19 +176,68 @@ class UnitTest(unittest.TestCase):
 
         self.assertEqual(self.gitr.new_files, self.new_list2)
 
+    @mock.patch("git_class.os")
     @mock.patch("git_class.gen_libs")
-    def test_process_newfiles_list(self, mock_lib):
+    def test_process_add_option(self, mock_lib, mock_os):
+
+        """Function:  test_process_add_option
+
+        Description:  Test with add option passed.
+
+        Arguments:
+
+        """
+
+        mock_lib.rm_file.return_value = True
+        mock_os.path.isdir.return_value = False
+
+        GIT = collections.namedtuple('GIT', 'index untracked_files')
+        DIFF = Diff()
+        self.gitr.gitrepo = GIT(DIFF, self.new_list1)
+        self.gitr.new_files = self.new_list2
+
+        self.gitr.process_untracked("add")
+
+        self.assertEqual(self.gitr.new_files, self.new_list2)
+
+    @mock.patch("git_class.os")
+    @mock.patch("git_class.gen_libs")
+    def test_process_remove_option(self, mock_lib, mock_os):
+
+        """Function:  test_process_remove_option
+
+        Description:  Test with remove option passed.
+
+        Arguments:
+
+        """
+
+        mock_lib.rm_file.return_value = True
+        mock_os.path.isdir.return_value = False
+
+        GIT = collections.namedtuple('GIT', 'index untracked_files')
+        DIFF = Diff()
+        self.gitr.gitrepo = GIT(DIFF, self.new_list1)
+        self.gitr.new_files = self.new_list2
+
+        self.gitr.process_untracked("remove")
+
+        self.assertEqual(self.gitr.new_files, self.new_list2)
+
+    @mock.patch("git_class.os")
+    @mock.patch("git_class.gen_libs")
+    def test_process_newfiles_list(self, mock_lib, mock_os):
 
         """Function:  test_process_newfiles_list
 
         Description:  Test with new_files list set.
 
         Arguments:
-            None
 
         """
 
         mock_lib.rm_file.return_value = True
+        mock_os.path.isdir.return_value = False
 
         GIT = collections.namedtuple('GIT', 'index untracked_files')
         DIFF = Diff()
@@ -227,19 +248,20 @@ class UnitTest(unittest.TestCase):
 
         self.assertEqual(self.gitr.new_files, self.new_list2)
 
+    @mock.patch("git_class.os")
     @mock.patch("git_class.gen_libs")
-    def test_process_empty_list2(self, mock_lib):
+    def test_process_empty_list2(self, mock_lib, mock_os):
 
         """Function:  test_process_empty_list2
 
         Description:  Test with empty list passed.
 
         Arguments:
-            None
 
         """
 
         mock_lib.rm_file.return_value = True
+        mock_os.path.isdir.return_value = False
 
         GIT = collections.namedtuple('GIT', 'index untracked_files')
         DIFF = Diff()
@@ -256,7 +278,6 @@ class UnitTest(unittest.TestCase):
         Description:  Test with empty list passed.
 
         Arguments:
-            None
 
         """